[MobileSyncExplorerSwift] Update to use split view and add visionOS support
