+++ MobileSyncExplorerSwift/MobileSyncExplorerSwift/SceneDelegate.swift
@@ -58,12 +58,15 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let url = URLContexts.first?.url, let userAccount = UserAccountManager.shared.currentUserAccount else {
             return
         }
+
         let sObjectManager = SObjectDataManager.sharedInstance(for: userAccount)
         if url.absoluteString.contains("contact/new") {
-            self.window?.rootViewController = UIHostingController(rootView: ContactListView(sObjectDataManager: sObjectManager, newContact: true))
+            self.window?.rootViewController = UIHostingController(rootView: ContactListView(sObjectManager: sObjectManager, newContact: true))
         } else if let contactRange = url.absoluteString.range(of: "contact/") {
             let id = String(url.absoluteString[contactRange.upperBound...])
-            self.window?.rootViewController = UIHostingController(rootView: ContactListView(sObjectDataManager: sObjectManager, selectedRecord: id))
+            self.window?.rootViewController = UIHostingController(rootView: NavigationStack {
+                ContactDetailView(localId: id, sObjectDataManager: sObjectManager)
+            })
         }
     }
     
@@ -86,13 +89,14 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
    
     func setupRootViewController(userActivity: NSUserActivity?) {
         let sObjectManager = SObjectDataManager.sharedInstance(for: UserAccountManager.shared.currentUserAccount!)
-        
+
         if let userActivity = userActivity, userActivity.title == openDetailPath,
            let selectionId = userActivity.userInfo?[openDetailRecordIdKey] as? String {
-            let list = ContactListView(sObjectDataManager: sObjectManager, selectedRecord: selectionId)
-            self.window?.rootViewController = UIHostingController(rootView: list)
+            self.window?.rootViewController = UIHostingController(rootView:  NavigationStack {
+                ContactDetailView(localId: selectionId, sObjectDataManager: sObjectManager)
+            })
         } else {
-            self.window?.rootViewController = UIHostingController(rootView: ContactListView(sObjectDataManager: sObjectManager))
+            self.window?.rootViewController = UIHostingController(rootView: Tabs(sObjectDataManager: sObjectManager))
         }
     }
