+++ MobileSyncExplorerSwift/RecentContacts/RecentContacts.swift
@@ -7,9 +7,12 @@
 //
 
 import Foundation
-import WidgetKit
 import UIKit.UIApplication
 import SalesforceSDKCore
+
+#if canImport(WidgetKit)
+import WidgetKit
+#endif
 
 struct Queue<Element: Codable> {
     var isEmpty: Bool {
@@ -84,7 +87,9 @@ class RecentContacts {
             let encryptionKey = try KeyGenerator.encryptionKey(for: encryptionKeyLabel)
             let encryptedData = try Encryptor.encrypt(data: encodedData, using: encryptionKey)
             try encryptedData.write(to: url)
-            WidgetCenter.shared.reloadAllTimelines()
+            #if canImport(WidgetKit)
+                WidgetCenter.shared.reloadAllTimelines()
+            #endif
         } catch {
             SalesforceLogger.e(RecentContacts.self, message: "Error persisting contacts: \(error)")
         }
